Description:

For srcdt==bf16 the current implementation of `jit_uni_eltwise_fwd_t` in oneDNN goes to ref implementation which is 10 times slower than f32 jit implementation. First noticed for eltwise_gelu_erf (benchmarked from PyTorch):

Gelu : [[1, 142, 3072]]
ref_implementation : 906 us
jit:sve_256 : 91 us


This patch invokes two reorder layers for up/down bf16/f32 casting in ideep and call the f32 jit eltwise primitive.
After this change, a 10x performance increase was noticed for shape : [1, 1500, 2048]:
Before :   5366 us
After :        556 us