@Ryo-not-rio 2025-01-07
Could you add some performance numbers to the PR message?